requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
# MCP 1.13.1 requires Pydantic >=2.11 and Uvicorn >=0.31.1
pydantic>=2.11,<3
mcp[cli]==1.13.1
httpx==0.27.2
anthropic==0.37.1
python-dotenv==1.0.1
langgraph==0.3.6
matplotlib==3.8.4
numpy==1.26.4